@Value 사용해 루트 디렉토리 경로 설정
# Conflicts:
#	src/main/java/com/sprint/mission/discodeit/controller/BinaryContentController.java
#	src/main/java/com/sprint/mission/discodeit/entity/MessageAttachment.java
#	src/main/java/com/sprint/mission/discodeit/service/BinaryContentService.java
#	src/main/java/com/sprint/mission/discodeit/service/BinaryContentStorage.java
#	src/main/java/com/sprint/mission/discodeit/service/FileManager.java
#	src/main/resources/application.yml
