+++ src/main/java/com/sprint/mission/discodeit/service/BinaryContentStorage.java
@@ -1,5 +1,7 @@
 package com.sprint.mission.discodeit.service;
 
+import org.springframework.core.io.UrlResource;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.io.InputStream;
@@ -9,5 +11,6 @@ public interface BinaryContentStorage {
 
     UUID put(UUID id, MultipartFile file);
     InputStream get(UUID id);
+    ResponseEntity<UrlResource> getUrl(UUID id);
     
 }
